ptr: tweak the `ptr` API a bit
This commit tweaks the `ptr` package API a bit. Its of course
very subjective and I'm happy to discuss (or close) the related
PR.

Concretely:
```go
-                       TLSVerify:    ptr.ToNilEmpty(container.TLSVerify),
+                       TLSVerify:    ptr.ToNilIfEmpty(container.TLSVerify),
```
here it feels like the intend is clearer with the "if" and it
(almost) adds no extra space.

```go
-                       InstallFrom:    ptr.FromOr(repo.Usage.Install, true),
+                       InstallFrom:    ptr.ValueOr(repo.Usage.Install, true),
```
Here I think either `Deref` or `Value` would be slightly quicker
to read.

```go
-                       Enabled:        ptr.Or(repo.Usage.Configure, true),
+                       Enabled:        ptr.OrTo(repo.Usage.Configure, true),
```
The `ptr.Or` feels a bit too terse (but I'm the least
sure about this one), conveying that this results in a pointer
to the `Or` value feels valuable to me (but maybe I'm overthinking
it).

I dropped `ptr.FromOrEmpty()` as I'm not seeing the difference
to `ptr.From()` (but maybe I'm just missing something).

I also dropped `ptr.ToErr()` as I feel its a bit magic and in the
only usage the diff has the same size.

I also dropped `ptr.IsZero()` as it is currently unused.