parser: make anyDetails more explicit
The current `anyDetails ...*AnyDetails` is a bit loose, ie.
it will allow to pass multiple anyDetails but only the first
is used. Just remove the `...` so that it becomes a static
compiler check (passing `nil` here is fine).

The alternative would be to do a runtime error if multiple
AnyDetails are passed but that static checks seems slightly
nicer (but its an option).