schema: remove pointers and use omitzero
This patch makes use of the new feature from oapi-codegen 2.5 named
prefer-skip-optional-pointer. When set to true, the default behavior is
to not generate pointer for struct fields, unless overridden in the
schema via x-go-type-skip-optional-pointer directly.

This means all the explicit OpenAPI schema fields are no longer
necessary and can be removed. There are few instances when it is needed:

* There is a bug in oapi-codegen and the global setting is not honoured
* Fields which we want to have pointer for (e.g. password)

Additionally, prefer-skip-optional-pointer-with-omitzero is specified to
generate omitzero for all fields which will cause the Go 1.24 JSON
parser to omit zero values (empty structs). This is only important for
convertors to prevent from generating structs with empty values.