schema: enable prefer-skip-optional-pointer
This patch makes use of the new feature from oapi-codegen 2.5 named
prefer-skip-optional-pointer. When set to true, the default behavior is
to not generate pointer for struct fields, unless overridden in the
schema via x-go-type-skip-optional-pointer directly.

This means all the explicit OpenAPI schema fields are no longer
necessary and can be removed. There are few instances when it is needed:

There is a bug in oapi-codegen and the global setting is not honoured
Fields which we want to have pointer for (e.g. password). These
instances are marked in the codebase.

Additionally, prefer-skip-optional-pointer-with-omitzero is disabled in
order for easeir review. It will be enabled later in the patch series so
a separate code is generated with only "omitzero" Go struct tags added.