deps: upgrade to Go 1.24
The JSON parser in Go 1.24 is the first version to support the "omitempty"
tag for struct fields. This vastly improves the output of the converter to JSON.
Without this upgrade, conversion of an empty UBP to BP will output something like:

    {
        "customizations": {
            "packages": {
            },
            "repositories": {
            },
            // ...
        }
    }

Note this is not needed for the loader, only for the convertor as a nice to have.