conv: use errors instead of logger for all problems for exporter
The exporter and importer types have a logger that is used for issuing warnings
and errors. The idea was to continue conversion even if some parts of the
blueprint were incorrect.

This patch changes that for the exporter to return errors instead of logging
them. The importer will be changed in a follow-up patch.