blueprint: new customization: dnf.config
New dnf.config customization block that enables one option for now:
set_releasever.  The option is meant to set the releasever dnf variable
to the value of the distro's release version, which ties a system to a
specific update of RHEL.  This prevents a system from being updated to a
later minor RHEL release through `dnf upgrade` and is the mechanism for
keeping systems on extended support releases (EUS and E4S).

The dnf customization block is created so that it might later be used
for other variables (perhaps even arbitrary, user-defined variables).
However, the releasever variable will be treated separately so that it
can be a simple switch that sets the `releasever` value to the one for
the image's distro dynamically.

The config key `set_releasever` was chosen over `set_release_ver`
because the variable name is `releasever` and it is more consistent with
other user-facing use cases where it is treated as a single word.